+++ CHANGES.rst
@@ -215,6 +215,14 @@ astropy.coordinates
 - Removed the ``recommended_units`` attribute from Representations; it was
   deprecated since 3.0. [#8892]
 
+- Removed the deprecated frame attribute classes, ``FrameAttribute``,
+  ``TimeFrameAttribute``,
+  ``QuantityFrameAttribute``,``CartesianRepresentationFrameAttribute``;
+  deprecated since 3.0. [#9326]
+
+- Removed ``longitude`` and ``latitude`` attributes from ``EarthLocation``;
+  deprecated since 2.0. [#9326]
+
 astropy.cosmology
 ^^^^^^^^^^^^^^^^^